+++ src/data_halfEdge.h
@@ -6,7 +6,7 @@
 // V 4
 // E 4
 // F 1
-float square_vertices[8] = {
+double square_vertices[8] = {
     0.5,
     0.5,
     0.5,
@@ -47,7 +47,7 @@ int square_hEdges[20] = {
 // V 4
 // E 4
 // F 1
-float two_square_vertices[12] = {
+double two_square_vertices[12] = {
     0.5,
     0.5,
     0.5,
